Split fonts into their own module
Just as there's a tones module, the fonts parsing stuff really doesn't
belong in "tools".
